requirements.txt
# Web Framework & API
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Data Processing & Analysis
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Background Jobs & Scheduling
apscheduler>=3.10.0

# Environment Variables
python-dotenv>=1.0.0

# Authentication & Security
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
slowapi>=0.1.9
